Moved folders and added Application
