+++ webapp/src/main/java/com/uniovi/wikidataservice/WichatWebappApplication.java
@@ -1,4 +1,4 @@
-package com.uniovi.wichatwebapp;
+package com.uniovi.wikidataservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
