+++ wikidataservice/src/main/java/com/uniovi/wikidataservice/ServletInitializer.java
@@ -1,0 +1,13 @@
+package com.uniovi.wikidataservice;
+
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+public class ServletInitializer extends SpringBootServletInitializer {
+
+    @Override
+    protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+        return application.sources(WikidataApplication.class);
+    }
+
+}
