+++ wikidataservice/src/main/java/com/uniovi/wikidataservice/WikidataApplication.java
@@ -1,0 +1,15 @@
+package com.uniovi.wikidataservice;
+
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WikidataApplication {
+    public WikidataApplication() { }
+
+    public static void main(java.lang.String[] args) {
+        SpringApplication.run(WikidataApplication.class, args);
+    }
+}
+
