+++ wikidataservice/src/main/java/com/uniovi/wikidataservice/entities/Question.java
@@ -1,9 +1,8 @@
-package com.uniovi.wichatwebapp.entities;
+package com.uniovi.wikidataservice.entities;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 import javax.persistence.*;
-import javax.swing.*;
 import java.util.ArrayList;
 import java.util.List;
