+++ wikidataservice/src/main/java/com/uniovi/wikidataservice/wikidata/WikidataUtils.java
@@ -1,4 +1,4 @@
-package com.uniovi.wichatwebapp.wikidata;
+package com.uniovi.wikidataservice.wikidata;
 
 public class WikidataUtils {
     public final static String LINKCONCAT = "#* &%";
