+++ wikidataservice/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+spring.application.name=wikidataservice
+
+server.port=8100
+
+spring.data.mongodb.uri=mongodb://localhost:27017/questions
